[ADD] add tab bar camera button
