+++ Puhaha/Puhaha/Assets.xcassets/EggFriedImage.imageset/EggFriedImage.pdf
@@ -1,0 +1,260 @@
+%PDF-1.7
+
+1 0 obj
+  << /ExtGState << /E1 << /ca 0.900000 >> >> >>
+endobj
+
+2 0 obj
+  << /Length 3 0 R >>
+stream
+/DeviceRGB CS
+/DeviceRGB cs
+q
+1.000000 0.000000 -0.000000 1.000000 0.000000 0.000000 cm
+1.000000 0.774011 0.000000 scn
+81.000000 40.500000 m
+81.000000 18.132469 62.867531 0.000000 40.500000 0.000000 c
+18.132467 0.000000 0.000000 18.132469 0.000000 40.500000 c
+0.000000 62.867531 18.132467 81.000000 40.500000 81.000000 c
+62.867531 81.000000 81.000000 62.867531 81.000000 40.500000 c
+h
+f
+n
+Q
+q
+q
+1.000000 0.000000 -0.000000 1.000000 15.299999 16.200195 cm
+1.000000 1.000000 1.000000 scn
+0.000000 35.500000 m
+0.000000 43.231987 6.268013 49.500000 13.999999 49.500000 c
+35.500000 49.500000 l
+43.231987 49.500000 49.500000 43.231987 49.500000 35.500000 c
+49.500000 14.000000 l
+49.500000 6.268013 43.231987 0.000000 35.500000 0.000000 c
+13.999999 0.000000 l
+6.268013 0.000000 0.000000 6.268013 0.000000 14.000000 c
+0.000000 35.500000 l
+h
+f
+n
+Q
+15.299999 51.700195 m
+15.299999 59.432182 21.568012 65.700195 29.299999 65.700195 c
+50.799999 65.700195 l
+58.531986 65.700195 64.800003 59.432182 64.800003 51.700195 c
+64.800003 30.200195 l
+64.800003 22.468208 58.531986 16.200195 50.799999 16.200195 c
+29.299999 16.200195 l
+21.568012 16.200195 15.299999 22.468208 15.299999 30.200195 c
+15.299999 51.700195 l
+h
+W*
+n
+q
+/E1 gs
+1.000000 0.000000 -0.000000 1.000000 15.299999 16.200195 cm
+1.000000 1.000000 1.000000 scn
+13.999999 46.500000 m
+35.500000 46.500000 l
+35.500000 52.500000 l
+13.999999 52.500000 l
+13.999999 46.500000 l
+h
+46.500000 35.500000 m
+46.500000 14.000000 l
+52.500000 14.000000 l
+52.500000 35.500000 l
+46.500000 35.500000 l
+h
+35.500000 3.000000 m
+13.999999 3.000000 l
+13.999999 -3.000000 l
+35.500000 -3.000000 l
+35.500000 3.000000 l
+h
+3.000000 14.000000 m
+3.000000 35.500000 l
+-3.000000 35.500000 l
+-3.000000 14.000000 l
+3.000000 14.000000 l
+h
+13.999999 3.000000 m
+7.924868 3.000000 3.000000 7.924866 3.000000 14.000000 c
+-3.000000 14.000000 l
+-3.000000 4.611160 4.611159 -3.000000 13.999999 -3.000000 c
+13.999999 3.000000 l
+h
+46.500000 14.000000 m
+46.500000 7.924866 41.575134 3.000000 35.500000 3.000000 c
+35.500000 -3.000000 l
+44.888840 -3.000000 52.500000 4.611160 52.500000 14.000000 c
+46.500000 14.000000 l
+h
+35.500000 46.500000 m
+41.575134 46.500000 46.500000 41.575134 46.500000 35.500000 c
+52.500000 35.500000 l
+52.500000 44.888840 44.888840 52.500000 35.500000 52.500000 c
+35.500000 46.500000 l
+h
+13.999999 52.500000 m
+4.611159 52.500000 -3.000000 44.888840 -3.000000 35.500000 c
+3.000000 35.500000 l
+3.000000 41.575134 7.924868 46.500000 13.999999 46.500000 c
+13.999999 52.500000 l
+h
+f
+n
+Q
+Q
+q
+q
+1.000000 0.000000 -0.000000 1.000000 29.324997 30.224609 cm
+1.000000 0.800376 0.116667 scn
+21.450001 10.725195 m
+21.450001 4.801941 16.648254 0.000195 10.725000 0.000195 c
+4.801746 0.000195 0.000000 4.801941 0.000000 10.725195 c
+0.000000 16.648449 4.801746 21.450195 10.725000 21.450195 c
+16.648254 21.450195 21.450001 16.648449 21.450001 10.725195 c
+h
+f
+n
+Q
+50.774998 40.949806 m
+50.774998 35.026550 45.973251 30.224804 40.049995 30.224804 c
+34.126743 30.224804 29.324997 35.026550 29.324997 40.949806 c
+29.324997 46.873058 34.126743 51.674805 40.049995 51.674805 c
+45.973251 51.674805 50.774998 46.873058 50.774998 40.949806 c
+h
+W*
+n
+q
+1.000000 0.000000 -0.000000 1.000000 29.324997 30.224609 cm
+1.000000 0.800376 0.116667 scn
+18.450001 10.725195 m
+18.450001 6.458795 14.991401 3.000195 10.725000 3.000195 c
+10.725000 -2.999805 l
+18.305109 -2.999805 24.450001 3.145086 24.450001 10.725195 c
+18.450001 10.725195 l
+h
+10.725000 3.000195 m
+6.458601 3.000195 3.000000 6.458795 3.000000 10.725195 c
+-3.000000 10.725195 l
+-3.000000 3.145086 3.144892 -2.999805 10.725000 -2.999805 c
+10.725000 3.000195 l
+h
+3.000000 10.725195 m
+3.000000 14.991594 6.458601 18.450195 10.725000 18.450195 c
+10.725000 24.450195 l
+3.144892 24.450195 -3.000000 18.305304 -3.000000 10.725195 c
+3.000000 10.725195 l
+h
+10.725000 18.450195 m
+14.991401 18.450195 18.450001 14.991594 18.450001 10.725195 c
+24.450001 10.725195 l
+24.450001 18.305304 18.305109 24.450195 10.725000 24.450195 c
+10.725000 18.450195 l
+h
+f
+n
+Q
+Q
+q
+q
+1.000000 0.000000 -0.000000 1.000000 54.551018 54.533203 cm
+1.000000 0.800376 0.116667 scn
+4.132653 1.997668 m
+4.132653 0.894507 3.207527 0.000219 2.066327 0.000219 c
+0.925126 0.000219 0.000000 0.894507 0.000000 1.997668 c
+0.000000 3.100829 0.925126 3.995117 2.066327 3.995117 c
+3.207527 3.995117 4.132653 3.100829 4.132653 1.997668 c
+h
+f
+n
+Q
+58.683670 56.530869 m
+58.683670 55.427711 57.758545 54.533424 56.617344 54.533424 c
+55.476143 54.533424 54.551018 55.427711 54.551018 56.530869 c
+54.551018 57.634033 55.476143 58.528320 56.617344 58.528320 c
+57.758545 58.528320 58.683670 57.634033 58.683670 56.530869 c
+h
+W*
+n
+q
+1.000000 0.000000 -0.000000 1.000000 54.551018 54.533203 cm
+1.000000 0.800376 0.116667 scn
+1.132653 1.997668 m
+1.132653 2.646415 1.647351 3.000219 2.066327 3.000219 c
+2.066327 -2.999781 l
+4.767704 -2.999781 7.132653 -0.857399 7.132653 1.997668 c
+1.132653 1.997668 l
+h
+2.066327 3.000219 m
+2.485303 3.000219 3.000000 2.646415 3.000000 1.997668 c
+-3.000000 1.997668 l
+-3.000000 -0.857399 -0.635051 -2.999781 2.066327 -2.999781 c
+2.066327 3.000219 l
+h
+3.000000 1.997668 m
+3.000000 1.348922 2.485303 0.995117 2.066327 0.995117 c
+2.066327 6.995117 l
+-0.635051 6.995117 -3.000000 4.852736 -3.000000 1.997668 c
+3.000000 1.997668 l
+h
+2.066327 0.995117 m
+1.647351 0.995117 1.132653 1.348922 1.132653 1.997668 c
+7.132653 1.997668 l
+7.132653 4.852736 4.767704 6.995117 2.066327 6.995117 c
+2.066327 0.995117 l
+h
+f
+n
+Q
+Q
+
+endstream
+endobj
+
+3 0 obj
+  5429
+endobj
+
+4 0 obj
+  << /Annots []
+     /Type /Page
+     /MediaBox [ 0.000000 0.000000 81.000000 81.000000 ]
+     /Resources 1 0 R
+     /Contents 2 0 R
+     /Parent 5 0 R
+  >>
+endobj
+
+5 0 obj
+  << /Kids [ 4 0 R ]
+     /Count 1
+     /Type /Pages
+  >>
+endobj
+
+6 0 obj
+  << /Pages 5 0 R
+     /Type /Catalog
+  >>
+endobj
+
+xref
+0 7
+0000000000 65535 f
+0000000010 00000 n
+0000000074 00000 n
+0000005559 00000 n
+0000005582 00000 n
+0000005755 00000 n
+0000005829 00000 n
+trailer
+<< /ID [ (some) (id) ]
+   /Root 6 0 R
+   /Size 7
+>>
+startxref
+5888
+%%EOF
